some changes on api services
- exception handling added in service
- new folder structure for api
